refactor(ts): move most of yargs-parser to lib/index.ts
